Refactoring items from collections from lists.
