+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/EventItem.kt
@@ -4,8 +4,6 @@ import android.os.CountDownTimer
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -22,44 +20,18 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.jordangellatly.hermosahappyhour.model.*
+import com.jordangellatly.hermosahappyhour.model.Event
+import com.jordangellatly.hermosahappyhour.model.RestaurantRepo
+import com.jordangellatly.hermosahappyhour.model.tower12HappyHour
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourCard
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 import java.text.SimpleDateFormat
 import java.util.*
 
 @Composable
-fun RestaurantCollection(
-    restaurants: List<Restaurant>,
-    onRestaurantClick: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    LazyColumn(
-        modifier = modifier,
-        contentPadding = PaddingValues(start = 12.dp, end = 12.dp)
-    ) {
-        items(restaurants) { restaurant ->
-            RestaurantItem(restaurant, onRestaurantClick)
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun RestaurantCollectionPreview() {
-    HermosaHappyHourTheme {
-        val restaurants = sampleSearchRestaurantData
-        RestaurantCollection(
-            restaurants = restaurants,
-            onRestaurantClick = {}
-        )
-    }
-}
-
-@Composable
-private fun RestaurantItem(
-    restaurant: Restaurant,
-    onRestaurantClick: (Long) -> Unit,
+fun EventItem(
+    event: Event,
+    onEventClick: (Long) -> Unit,
     modifier: Modifier = Modifier
 ) {
     HappyHourCard(
@@ -67,17 +39,13 @@ private fun RestaurantItem(
             .fillMaxWidth()
             .padding(bottom = 16.dp)
     ) {
-        val date = getCurrentDateTime()
-        val getDayFromDate = date.toString("EEEE").uppercase()
-        val happyHourToday =
-            restaurant.eventsToday.getValue(getDayFromDate).getValue(EventType.HappyHour)
-        val happyHours = happyHourToday.hours
+        val eventHours = event.hours
 
         val currentTime = Calendar.getInstance()
         val startTime = Calendar.getInstance()
         val endTime = Calendar.getInstance()
 
-        val splitStartTimeFromHours = happyHours.split("-")
+        val splitStartTimeFromHours = eventHours.split("-")
         val stringStart = splitStartTimeFromHours[0].trim()
         val stringEnd = splitStartTimeFromHours.get(index = 1).trim()
 
@@ -101,9 +69,10 @@ private fun RestaurantItem(
 
         Column(
             modifier = Modifier
-                .clickable(onClick = { onRestaurantClick(restaurant.id) })
+                .clickable(onClick = { onEventClick(event.restaurantId) })
                 .fillMaxSize()
         ) {
+            val restaurant = remember { RestaurantRepo.getRestaurant(event.restaurantId) }
             Box(
                 modifier = Modifier
                     .height(160.dp)
@@ -130,7 +99,7 @@ private fun RestaurantItem(
             var millisInFuture = 0L
             var timeIndicatorColor = HermosaHappyHourTheme.colors.textSecondary
             val annotatedTimeString = buildAnnotatedString {
-                append("Happy Hour \u2022 ")
+                append("${event.title} \u2022 ")
                 when {
                     currentTime < startTime -> {
                         timeIndicatorColor = Color.Green
@@ -196,14 +165,14 @@ private fun RestaurantItem(
     }
 }
 
-@Preview
+@Preview(heightDp = 240)
 @Composable
-private fun RestaurantItemPreview() {
+private fun EventItemPreview() {
     HermosaHappyHourTheme {
-        val restaurant = tower12
-        RestaurantItem(
-            restaurant = restaurant,
-            onRestaurantClick = {}
+        val event = tower12HappyHour
+        EventItem(
+            event = event,
+            onEventClick = {}
         )
     }
 }
