+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/search/SearchFeed.kt
@@ -1,6 +1,8 @@
-package com.jordangellatly.hermosahappyhour.ui.home
+package com.jordangellatly.hermosahappyhour.ui.search
 
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
@@ -10,32 +12,33 @@ import com.jordangellatly.hermosahappyhour.model.EventType
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.RestaurantRepo
 import com.jordangellatly.hermosahappyhour.ui.components.HermosaHappyHourSurface
+import com.jordangellatly.hermosahappyhour.ui.home.DateBar
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 
 @Composable
-fun RestaurantFeed(
+fun SearchFeed(
     onRestaurantClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
     eventType: EventType = EventType.HappyHour
 ) {
     val restaurantList = remember { RestaurantRepo.getRestaurantsByEventType(eventType) }
-    RestaurantFeed(
-        restaurantList = restaurantList,
+    SearchFeed(
+        restaurants = restaurantList,
         onRestaurantClick = onRestaurantClick,
         modifier = modifier
     )
 }
 
 @Composable
-private fun RestaurantFeed(
-    restaurantList: List<Restaurant>,
+private fun SearchFeed(
+    restaurants: List<Restaurant>,
     onRestaurantClick: (Long) -> Unit,
     modifier: Modifier = Modifier
 ) {
     HermosaHappyHourSurface(modifier = modifier.fillMaxSize()) {
         Box {
             RestaurantList(
-                restaurantList = restaurantList,
+                restaurants = restaurants,
                 onRestaurantClick = onRestaurantClick
             )
             DateBar()
@@ -45,7 +48,7 @@ private fun RestaurantFeed(
 
 @Composable
 private fun RestaurantList(
-    restaurantList: List<Restaurant>,
+    restaurants: List<Restaurant>,
     onRestaurantClick: (Long) -> Unit,
     modifier: Modifier = Modifier
 ) {
@@ -56,10 +59,14 @@ private fun RestaurantList(
                     WindowInsets.statusBars.add(WindowInsets(top = 56.dp))
                 )
             )
-            RestaurantCollection(
-                restaurants = restaurantList,
-                onRestaurantClick = onRestaurantClick
-            )
+            LazyColumn(
+                modifier = modifier,
+                contentPadding = PaddingValues(start = 12.dp, end = 12.dp)
+            ) {
+                items(restaurants) { restaurant ->
+                    RestaurantItem(restaurant, onRestaurantClick)
+                }
+            }
         }
     }
 }
@@ -68,6 +75,6 @@ private fun RestaurantList(
 @Composable
 private fun RestaurantFeedPreview() {
     HermosaHappyHourTheme {
-        RestaurantFeed(onRestaurantClick = { })
+        SearchFeed(onRestaurantClick = { })
     }
 }
